Adden new parameter to fmin
